Move log to seperate file
